+++ apps/webportal/src/pages/department/[id]/class-routine/[routineSlug].jsx
@@ -1,0 +1,12 @@
+import React, { Fragment } from 'react';
+import ClassRoutineTable from '../../../../components/ClassRoutineTable';
+
+const Routine = () => {
+  return (
+    <Fragment>
+      <ClassRoutineTable />
+    </Fragment>
+  );
+};
+
+export default Routine;
