+++ apps/webportal/src/pages/department/[id]/course/[courseId]/offer-and-distribute/index.tsx
@@ -10,7 +10,6 @@ import {
   Grid,
   InputLabel,
   MenuItem,
-  Paper,
   Select,
   TextField,
 } from '@mui/material';
