+++ apps/webportal/src/styles/globals.css
@@ -4,3 +4,20 @@
   text-decoration: none;
   color: inherit;
 }
+::-webkit-scrollbar {
+  width: 10px;
+  height: 13px;
+}
+
+::-webkit-scrollbar-thumb {
+  background-clip: padding-box;
+  border-radius: 10px;
+  background-color: rgb(8 19 72 / 30%);
+  opacity: 0.3;
+}
+
+/* Track */
+::-webkit-scrollbar-track {
+  background-color: #e2e6f5;
+  border-radius: 10px;
+}
